refactor: update UI components to use SuperSectionCard2 for better layout and consistency across settings and store screens
- Replaced MCard with SuperSectionCard2 in IntegrationsView, NotificationsView, RoleEditorView, TaxesSettingsView, WebhooksView, CreateStoreScreen, IssueInventoryScreen, and StoreDetailScreen.
- Improved the structure and readability of the code by organizing child widgets within SuperSectionCard2.
- Updated pubspec.yaml to use super_core version 3.1.0 for enhanced features and stability.

+++ lib/features/mobile_dashboard/presentation/widgets/mobile_dashboard_section_view.dart
@@ -387,7 +387,7 @@ class _MobileDashboardSectionViewState
                               MobileDashboardQuickActionsSkeleton(
                                 itemCount: selectedTab.actions.isEmpty
                                     ? 8
-                                    : (selectedTab.actions.length.clamp(4, 8) as int),
+                                    : selectedTab.actions.length.clamp(4, 8),
                               )
                             else
                               MobileDashboardQuickActions(
